Main/requirements.txt
ultralytics
opencv-python
numpy
cupy
torch
torchvision
torchaudio
torch-directml
onnx
onnx-simplifier
onnxruntime
onnxruntime-directml
pycuda
tensorrt
colorama
customtkinter
pywin32
pandas
requests
